refactor(plugins/hooks-extra): improve code reusability
